+++ test/Grpc.IntegrationTests/GrpcDurableTaskClientIntegrationTests.cs
@@ -101,7 +101,7 @@ public class DurableTaskGrpcClientIntegrationTests : IntegrationTestBase
         await ForEachOrchestrationAsync(
             x => server.Client.ScheduleNewOrchestrationInstanceAsync(
                 OrchestrationName, input: false, new StartOrchestrationOptions(x)));
-        AsyncPageable<OrchestrationMetadata> pageable = server.Client.GetInstancesAsync(query);
+        AsyncPageable<OrchestrationMetadata> pageable = server.Client.GetAllInstancesAsync(query);
 
         await ForEachOrchestrationAsync(x => server.Client.WaitForInstanceStartAsync(x, default));
         List<OrchestrationMetadata> metadata = await pageable.ToListAsync();
@@ -129,7 +129,7 @@ public class DurableTaskGrpcClientIntegrationTests : IntegrationTestBase
                 OrchestrationName, input: false, new StartOrchestrationOptions($"GetInstances_AsPages_EndToEnd-{i}"));
         }
 
-        AsyncPageable<OrchestrationMetadata> pageable = server.Client.GetInstancesAsync(query);
+        AsyncPageable<OrchestrationMetadata> pageable = server.Client.GetAllInstancesAsync(query);
         List<Page<OrchestrationMetadata>> pages = await pageable.AsPages(pageSizeHint: 5).ToListAsync();
         pages.Should().HaveCount(5);
         pages.ForEach(p => p.Values.Should().HaveCount(p.ContinuationToken is null ? 1 : 5));