+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `DurableTaskClient` methods have been touched up to ensure `CancellationToken` is included, as well as is the last parameter.
 - Removed obsolete/unimplemented local lambda activity calls from `TaskOrchestrationContext`
 - Input is now an optional parameter on `TaskOrchestrationContext.ContinueAsNew`
+- Multi-target gRPC projects to now use `Grpc.Net.Client` when appropriate (.NET6.0 and up)
 
 ## v1.0.0-rc.1