+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ﻿# Changelog
 
+## v1.0.1
+
+### Microsoft.DurableTask.Client
+
+- Fix incorrect bounds check on `PurgeResult`
+- Address typo for `DurableTaskClient.GetInstancesAsync` (incorrectly pluralized)
+    - Added `GetInstanceAsync`
+    - Hide `GetInstancesAsync` from editor
+
 ## v1.0.0
 
 - Added `SuspendInstanceAsync` and `ResumeInstanceAsync` to `DurableTaskClient`.